Big changes! Add variable information recording. Refactor SymbolTable to take in tokens and translate into variables internally, add appropriate calls to track variable usage in parser, change around fields of Variable / SymbolTable a bit.

+++ src/parser/scanner.rs
@@ -7,7 +7,7 @@ pub struct Scanner {
     contents : String,
     curr_ch  : Option<char>,
     position : usize,
-    line     : u16,
+    line     : usize,
     put_back : bool,
 }
